+++ src/main/java/ru/bellintegrator/practice/organization/service/impl/OrganizationServiceImpl.java
@@ -7,6 +7,8 @@ import org.springframework.context.annotation.Scope;
 import org.springframework.context.annotation.ScopedProxyMode;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.bellintegrator.practice.common.exception.EntityDoesNotExistException;
+import ru.bellintegrator.practice.common.exception.OrgDoesNotExistException;
 import ru.bellintegrator.practice.office.dao.OfficeDAO;
 import ru.bellintegrator.practice.office.model.Office;
 import ru.bellintegrator.practice.organization.dao.OrganizationDAO;
@@ -40,9 +42,14 @@ public class OrganizationServiceImpl implements OrganizationService {
     public List<OrganizationFilterOutView> filterByName(OrganizationFilterInView inView) {
         Organization org = new Organization();
 
+        //Set Org name
         org.setOrgName(inView.getName());
-        org.setInn(inView.getInn());
-        org.setActive(inView.getActive());
+        //Set Org inn
+        if ((inView.getInn() != null) && (!inView.getInn().isEmpty()))
+            org.setInn(inView.getInn());
+        //Set Org isActive
+        if ((inView.getActive() != null) && (!inView.getActive().isEmpty()))
+            org.setActive(Boolean.parseBoolean(inView.getActive()));
 
         List<Organization> all = dao.filterByName(org);
         List<OrganizationFilterOutView> outViews = new ArrayList<>();
@@ -58,7 +65,7 @@ public class OrganizationServiceImpl implements OrganizationService {
             List<Office> allOffices = officeDAO.filterByOrgId(office);
 
             OrganizationFilterOutView outView = new OrganizationFilterOutView(orgLoop.getId().toString(),
-                    orgLoop.getOrgName(), String.valueOf(allOffices.size()), orgLoop.getActive());
+                    orgLoop.getOrgName(), String.valueOf(allOffices.size()), orgLoop.getActive().toString());
             outViews.add(outView);
 
             log.info(outView.toString());
@@ -73,8 +80,12 @@ public class OrganizationServiceImpl implements OrganizationService {
     public OrganizationView loadById(String id) {
         Organization org = dao.loadById(Long.parseLong(id));
 
+        //Checks if Org exist
+        if (org == null)
+            throw new EntityDoesNotExistException("Org", id);
+
         OrganizationView view = new OrganizationView(org.getId().toString(), org.getOrgName(), org.getOrgFullName(),
-                org.getInn(), org.getKpp(), org.getAddress(), org.getPhoneNumber(), org.getActive());
+                org.getInn(), org.getKpp(), org.getAddress(), org.getPhoneNumber(), org.getActive().toString());
 
         log.info(view.toString());
 
@@ -85,6 +96,10 @@ public class OrganizationServiceImpl implements OrganizationService {
     @Transactional
     public void updateById(OrganizationView inView) {
         Organization org = dao.loadById(Long.parseLong(inView.getId()));
+
+        //Checks if Org exist
+        if (org == null)
+            throw new EntityDoesNotExistException("Org", inView.getId());
 
         //Org name
         if ((inView.getName() != null) && (!inView.getName().isEmpty()))
@@ -105,8 +120,8 @@ public class OrganizationServiceImpl implements OrganizationService {
         if ((inView.getPhone() != null) && (!inView.getPhone().isEmpty()))
             org.setPhoneNumber(inView.getPhone());
         //Is Org active
-        if (inView.getActive() != null)
-            org.setActive(inView.getActive());
+        if ((inView.getActive() != null) && (!inView.getActive().isEmpty()))
+            org.setActive(Boolean.parseBoolean(inView.getActive()));
     }
 
     @Override
@@ -114,6 +129,7 @@ public class OrganizationServiceImpl implements OrganizationService {
     public OrganizationIdOutView save(OrganizationSaveView inView) {
         Organization org = new Organization();
 
+        //Obligatory part
         //Org name
         org.setOrgName(inView.getName());
         //Org full name
@@ -122,15 +138,20 @@ public class OrganizationServiceImpl implements OrganizationService {
         org.setInn(inView.getInn());
         //kpp
         org.setKpp(inView.getKpp());
-        //address
+        //Org address
         org.setAddress(inView.getAddress());
-        //phone
-        org.setPhoneNumber(inView.getPhone());
+
+        //Optional part
+        //Org phone
+        if ((inView.getPhone() != null) && (!inView.getPhone().isEmpty()))
+            org.setPhoneNumber(inView.getPhone());
         //Is Org active
-        if (inView.getActive() != null)
-            org.setActive(inView.getActive());
-        else
+        if ((inView.getActive() != null) && (!inView.getActive().isEmpty()))
+            org.setActive(Boolean.parseBoolean(inView.getActive()));
+        else {
             org.setActive(true);
+            inView.setActive("true");
+        }
 
         log.info(inView.toString());
 
@@ -145,6 +166,11 @@ public class OrganizationServiceImpl implements OrganizationService {
     @Transactional
     public void deleteById(String id) {
         Organization org = dao.loadById(Long.parseLong(id));
+
+        //Checks if Org exist
+        if (org == null)
+            throw new EntityDoesNotExistException("Org", id);
+
         dao.deleteById(org);
     }
 }
